hypertrend-backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asyncpg==0.31.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.1.15
django-admin==2.0.2
django-cors-headers==4.9.0
django-excel-response2==3.0.6
django-haystack==3.3.0
django-six==1.0.5
djangorestframework==3.15.2
drf-haystack==1.9.1
excel-base==1.0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isoweek==1.3.3
mthrottle==0.0.2
nse==1.2.9
numpy==2.3.5
packaging==25.0
pandas==2.3.3
psycopg2==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.1.0
requests==2.32.5
screen==1.0.1
six==1.17.0
sqlparse==0.5.4
TimeConvert==3.0.13
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
xlwt==1.3.0